Flask/requirements.txt
allure-pytest==2.14.3
allure-python-commons==2.14.3
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
gunicorn==23.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
psycopg2==2.9.10
Pygments==2.19.2
pytest==8.4.1
requests==2.31.0
selenium==4.34.2
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.14.1
Werkzeug==3.1.3
